fix 'return_direct' and unify env helper
our ssh tools had 'returns_direct' set to True which means that the
agent loop should stop when the tool-call si finished. This was
(for some reason) mostly ignored by `initial_version.py` but let
the `react` version to stop after the first tool invocation.

By fixing this, the `initial version` also beccomes more clearer (as
in the version before multiple tool invocations looked as if they
happened in parallel)